+++ Ch06/06_02/start/react-app/src/App.js
@@ -1,6 +1,16 @@
 import "./App.css";
 import { useState, useEffect } from "react";
 
+function GithubUser({name, location, avatar}) {
+  return (
+    <div>
+      <h1>{name}</h1>
+      <p>{location}</p>
+      <img src={avatar} height={150} alt={name}/>
+    </div>
+  )
+}
+
 function App() {
   const [data, setData] = useState(null);
   useEffect(() => {
@@ -10,9 +20,14 @@ function App() {
       .then((response) => response.json())
       .then(setData);
   }, []);
+
   if (data)
     return (
-      <pre>{JSON.stringify(data, null, 2)}</pre>
+      <GithubUser 
+        name={data.name}
+        location={data.location} 
+        avatar={data.avatar_url}
+      />
     );
   return <h1>Data</h1>;
 }
