+++ Ch06/06_03/start/react-app/src/App.js
@@ -13,22 +13,31 @@ function GithubUser({ name, location, avatar }) {
 
 function App() {
   const [data, setData] = useState(null);
+  const [error, setError] = useState(null);
+  const [loading, setLoading] = useState(false);  // handling loading state
+
   useEffect(() => {
+    setLoading(true);
     fetch(
       `https://api.github.com/users/moonhighway`
     )
       .then((response) => response.json())
-      .then(setData);
+      .then(setData)
+      .then(() => setLoading(false))
+      .catch(setError);
   }, []);
-  if (data)
-    return (
-      <GithubUser
-        name={data.name}
-        location={data.location}
-        avatar={data.avatar_url}
-      />
-    );
-  return <h1>Data</h1>;
+
+  if (loading) return <h1>Loading...</h1>;
+  if (error) 
+    return <pre>{JSON.stringify(error)}</pre>;
+  if (!data) return null;
+  return (
+    <GithubUser
+      name={data.name}
+      location={data.location}
+      avatar={data.avatar_url}
+    />
+  );
 }
 
 export default App;
