+++ Ch06/06_05/start/react-app/src/App.js
@@ -1,64 +1,36 @@
 import "./App.css";
-import { useState, useEffect } from "react";
 
-const query = `
-query {
-  allLifts {
-    name
-    elevationGain
-    status
-  }
-}
-`;
+const tahoe_peaks = [
+  { name: "Freel", elevation: 10891},
+  { name: "Monument", elevation: 10067},
+  { name: "Pyramid", elevation: 9989},
+  { name: "Tallac", elevation: 9735}
+];
 
-const opts = {
-  method: "POST",
-  headers: { "Content-Type": "application/json" },
-  body: JSON.stringify({ query })
-};
-
-function Lift({ name, elevationGain, status }) {
-  return (
-    <div>
-      <h1>{name}</h1>
-      <p>
-        {elevationGain} {status}
-      </p>
-    </div>
+// if data does not exist, return renderEmpty
+//else return <ul></ul>
+function List({data, renderItem, renderEmpty}) {
+  return !data.length ? (
+    renderEmpty
+  ) : (
+    <ul>
+      {data.map((item) => (
+        <li key={item.name}>{renderItem(item)}</li>
+      ))}
+    </ul>
   );
 }
 
 function App() {
-  const [data, setData] = useState(null);
-  const [error, setError] = useState(null);
-  const [loading, setLoading] = useState(false);
-
-  useEffect(() => {
-    setLoading(true);
-    fetch(
-      `https://snowtooth.moonhighway.com/`,
-      opts
-    )
-      .then((response) => response.json())
-      .then(setData)
-      .then(() => setLoading(false))
-      .catch(setError);
-  }, []);
-
-  if (loading) return <h1>Loading...</h1>;
-  if (error)
-    return <pre>{JSON.stringify(error)}</pre>;
-  if (!data) return null;
   return (
-    <div>
-      {data.data.allLifts.map((lift) => (
-        <Lift
-          name={lift.name}
-          elevationGain={lift.elevationGain}
-          status={lift.status}
-        />
-      ))}
-    </div>
+    <List
+      data={tahoe_peaks}
+      renderItem={item => (
+        // <></> react fragment
+        <>{item.name} - {item.elevation} ft.</>
+      )}
+      renderEmpty={<p>This list is empty</p>}
+    />
   );
 }
